experimenting with different frequency eligibility criteria
